+++ sources/client/udpClientSocket/udpClientSocket.hpp
@@ -29,7 +29,7 @@ class udpClientSocket {
          * @param t_udpPort The UDP port to use.
          * @param t_ip The IP address to use.
          */
-        udpClientSocket(size_t t_udpPort, ip::address t_ip);
+        udpClientSocket(unsigned int t_udpPort, ip::address t_ip);
 
         /**
          * @brief Destroy the udpClientSocket object.
